remove eslint flag
