+++ src/index.js
@@ -2,9 +2,12 @@ import createList from './createList.js';
 import create from './localStorage.js';
 import './index.scss';
 
+const todoList = document.querySelector('.todo-list');
 const input = document.getElementById('input');
 const clearButton = document.querySelector('.clear');
+
 let output = [];
+
 input.addEventListener('keydown', (event) => {
   if (event.key === 'Enter' && input.value !== '') {
     create(input.value, false, output);
@@ -12,6 +15,34 @@ input.addEventListener('keydown', (event) => {
     input.value = '';
   }
 });
+
+//* Delete a selected list
+
+export const deleteList = (list, deleteToggle, index) => {
+  output = []
+  deleteToggle.addEventListener('click', () => {
+    todoList.removeChild(list);
+
+    const data = JSON.parse(localStorage.getItem('data'));
+
+    data.forEach((element) => {
+      if (element.index === index) {
+        data.splice(index, 1);
+      }
+    });
+    data.forEach((item, index) => {
+      item.index = index;
+      output.push(item)
+    });
+
+    localStorage.setItem('data', JSON.stringify(data));
+
+    createList();
+  });
+
+  list.style.backgroundColor = '#FFFEC3';
+};
+
 
 const clear = () => {
   output = [];
